number of unique terms may not fit in int
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1074880 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/util/automaton/fst/Builder.java
@@ -91,7 +91,7 @@ public class Builder<T> {
     return fst.nodeCount;
   }
 
-  public int getTermCount() {
+  public long getTermCount() {
     return frontier[0].inputCount;
   }
 
@@ -422,7 +422,7 @@ public class Builder<T> {
     Arc<T>[] arcs;
     T output;
     boolean isFinal;
-    int inputCount;
+    long inputCount;
 
     /** This node's depth, starting from the automaton root. */
     final int depth;
